requirements.txt
Django==3.1
django-admin-sortable2==0.7.7
django-tinymce==3.0.2
Pillow==7.1.2
python-dotenv==0.12.0
requests==2.23.0